move lib into src
