+++ src/components/game/PhasorGame.tsx
@@ -17,7 +17,7 @@ export const PhaserGame = forwardRef<IRefPhaserGame>(
     useLayoutEffect(() => {
       if (gameRef.current === null) {
         const loadGame = async () => {
-          const { default: StartGame } = await import("src/phaser/main");
+          const { default: StartGame } = await import("phaser/main");
           gameRef.current = StartGame(containerId);
         };
         loadGame();
